TST/MNT/BLD Lots of changes
TST: Add a test to the io functionality. Load a spec file and try
to plot one of the columns. Then check to make sure that the returned
mpl artist has more than 0 data points

BLD: Package a sample spec data set

MNT: Fix the plotting in the io module